+++ godemon.go
@@ -1,4 +1,4 @@
-package main
+package godemon
 
 import (
 	"encoding/json"
@@ -6,16 +6,17 @@ import (
 	"io"
 	"log"
 	"os"
-	"os/exec"
 	"path/filepath"
+	"time"
 )
 
-func BLR(path string) (*FileTreeNode, error) {
+func BLR(path string, t time.Time) (*FileTreeNode, error) {
+
 	n := newFileNode(path)
 	if n.Value.IsDir() {
 		files, err := os.ReadDir(path)
 		check(err)
-		dirignore := make(map[string]bool, 0)
+		dirignore := make(map[string]bool)
 		err = IgnoreDirs(dirignore)
 		check(err)
 		for _, f := range files {
@@ -24,9 +25,13 @@ func BLR(path string) (*FileTreeNode, error) {
 			}
 			childPath := filepath.Join(path, f.Name())
 			childNode := newFileNode(childPath)
-			fmt.Printf("name: %v\n path: %v\n modTime: %v\n", f.Name(), childNode.Path, childNode.Value.ModTime())
 
-			_, _ = BLR(childPath)
+			_, _ = BLR(childPath, childNode.Value.ModTime())
+			fmt.Printf("name: %v\n path: %v\n modTime: %v\n", f.Name(), childNode.Path, childNode.Value.ModTime())
+			if !childNode.Value.ModTime().Equal(t) {
+				log.Printf("%v: %v", ErrChanged, childPath)
+				return childNode, ErrChanged
+			}
 			if childNode != nil {
 				n.Children = append(n.Children, childNode)
 			}
@@ -44,7 +49,7 @@ func IgnoreDirs(ignoreDirs map[string]bool) error {
 
 	b, err := io.ReadAll(jsonF)
 	if err != nil {
-		return fmt.Errorf("error", err)
+		return fmt.Errorf("error: %v", err)
 	}
 	err = json.Unmarshal(b, &ignoreDirs)
 	if err != nil {
@@ -53,62 +58,30 @@ func IgnoreDirs(ignoreDirs map[string]bool) error {
 	return nil
 }
 
-//func (root *FileTreeNode) GodemonInit() error {
-//  for {
-//    initTime := time.Now()
-//    root.BFS(initTime)
-//    // logic for restart and ongoing
-//    if root.Error() != nil {
-//      fmt.Printf("There is an error in programme: %v\n", root.Error())
-//      return root.Error()
-//    }
-//  }
+//func GodemonInit() error {
+//	prevTime := time.Now()
+//	for {
+//		root, err := BLR(".", prevTime)
+//		if checkf("There is some trouble with app", err) {
+//			switch {
+//			case err == ErrChanged:
+//
+//			}
+//		}
+//
+//		time.Sleep(400 * time.Millisecond)
+//	}
 //}
-
-func shell() error {
-	absPath, err := filepath.Abs(".")
-	check(err)
-	rootName, err := os.Stat(absPath)
-	check(err)
-	scriptName := "restart.bat"
-	// Define the contents of the batch file
-	contents := fmt.Sprintf(`
-		echo hello world
-		go build %v
-		%v\%v.exe
-	`, absPath, absPath, rootName)
-
-	// Create the batch file
-	file, err := os.Create("restart.bat")
-	if err != nil {
-		log.Println("there is an error: ", err)
-		return err
-	}
-	defer file.Close()
-
-	// Write the contents to the file
-	_, err = file.WriteString(contents)
-	if err != nil {
-		panic(err)
-	}
-
-	// Execute the batch file
-	err = exec.Command(filepath.Join(absPath, scriptName)).Run()
-	if err != nil {
-		fmt.Printf("Failed to execute restart.bat: %v\n", err)
-		return err
-	}
-	fmt.Println("Server Restarted")
-	return nil
-}
-
-func main() {
-	thing, err := BLR(".")
-	check(err)
-	fmt.Println(thing.Path, thing.Value.ModTime())
-
-	for i := 0; i < len(thing.Children); i++ {
-		fmt.Println(thing.Children[i].Value.Name())
-	}
-
-}
+//
+//func Main() {
+//	prevTime := time.Now()
+//	thing, err := BLR(".", prevTime)
+//	check(err)
+//	fmt.Println(thing.Path, thing.Value.ModTime())
+//
+//	for i := 0; i < len(thing.Children); i++ {
+//		fmt.Println(thing.Children[i].Value.Name())
+//	}
+//	err = shell()
+//	check(err)
+//}
